Added calculate stat methods to IStat
Moved calculations from Stat class to BattleStat and PermanentStat enums

+++ src/main/java/com/github/drsmugbrain/pokemon/Pokemon.java
@@ -113,7 +113,7 @@ public class Pokemon implements ITypes {
         }
         STATS = new Stats(builder);
 
-        hp = PermanentStat.HP.calculate(this);
+        hp = (int) PermanentStat.HP.calculate(this);
 
         MOVES.addAll(moves);
         weight = basePokemon.getWeight();
